requirements.txt
colorama
matplotlib
numpy
torch
torchvision
tqdm
opencv-python
pygame
Pillow